Done Task 2

+++ home.php
@@ -1,7 +1,7 @@
-<?php 
-    include 'action.php';
-    require_once "config.php";
-    // session_start();
+<?php
+include 'action.php';
+require_once "config.php";
+// session_start();
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -35,15 +35,15 @@
     </button>
     <div class="collapse navbar-collapse" id="navbarNav">
         <ul class="navbar-nav ml-auto">
-        <li class="nav-item active">
-            <a class="nav-link" href="home">Project Management<span class="sr-only">(current)</span></a>
-        </li>
-        <li class="nav-item">
-            <a class="nav-link" href="bigquery">Project BigQuery</a>
-        </li>
-        <li class="nav-item">
-            <a class="nav-link" href="#">Additional App</a>
-        </li>
+            <li class="nav-item active">
+                <a class="nav-link" href="home">Project Management<span class="sr-only">(current)</span></a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="bigquery">Project BigQuery</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="#">Additional App</a>
+            </li>
         </ul>
     </div>
 </nav>
@@ -58,7 +58,7 @@
         <a class="btn btn-success" href="form">Add New Project
             <i class="fa fa-plus" aria-hidden="true"></i>
         </a>
-            
+
         <a class="btn btn-secondary" href="#bottom">Jump to bottom</a>
 
         <!-- Table -->
@@ -78,31 +78,31 @@
                 </tr>
             </thead>
             <tbody>
-                </tr>  
-                    <?php //  Only show the required information which are: ID, Project Name, Subtype, Current Status,
-                    // Country, Province/State, District, Latitude, Longitude
-                    $mainInfoCol = array(0,1,2,3,12,13,14,16,17);
-                    $file = fopen($FILE_PATH, 'r');
-                    while (($line = fgetcsv($file)) !== FALSE):
-                        echo "<tr>";
-                        foreach($mainInfoCol as $col) {
-                            echo "<td>$line[$col]</td>";
-                        } ?>
-                        <td>
-                            <a href="form?view=<?=$line[0];?>" class="btn btn-primary btn-sm">Details/Update
-                                <i class="fa fa-pencil-square-o" aria-hidden="true"></i>
-                            </a>
-                            <a href="action.php?delete=<?=$line[0];?>" class="btn btn-danger btn-sm"
-                            onclick="return confirm('Do you want to delete this record?');">Delete
-                                <i class="fa fa-trash" aria-hidden="true"></i>
-                            </a>
-                        </td>
-                        <?php  endwhile; fclose($file); ?>
                 </tr>
-            </tbody> 
+                <?php //  Only show the required information which are: ID, Project Name, Subtype, Current Status,
+                // Country, Province/State, District, Latitude, Longitude
+                $mainInfoCol = array(0, 1, 2, 3, 12, 13, 14, 16, 17);
+                $file = fopen($FILE_PATH, 'r');
+                while (($line = fgetcsv($file)) !== FALSE) :
+                    echo "<tr>";
+                    foreach ($mainInfoCol as $col) {
+                        echo "<td>$line[$col]</td>";
+                    } ?>
+                    <td>
+                        <a href="form?view=<?= $line[0]; ?>" class="btn btn-primary btn-sm">Details/Update
+                            <i class="fa fa-pencil-square-o" aria-hidden="true"></i>
+                        </a>
+                        <a href="action.php?delete=<?= $line[0]; ?>" class="btn btn-danger btn-sm" onclick="return confirm('Do you want to delete this record?');">Delete
+                            <i class="fa fa-trash" aria-hidden="true"></i>
+                        </a>
+                    </td>
+                <?php endwhile;
+                fclose($file); ?>
+                </tr>
+            </tbody>
         </table>
         <label id="bottom">End of page</label>
-        <a class="btn btn-secondary" href="#top">Jump to top</a>        
+        <a class="btn btn-secondary" href="#top">Jump to top</a>
     </div>
 </body>
